Fix for bug #869
This fix saves and restores the generated names for program variables. It uses the same technique as the fix for anonymous updates, it gives proposals to VariableNamer and saves the generated names in a SchemaVariable.
Also, from now on new program variables which are added to GlobalProgVars are added to NamespaceSet.programVariables() to avoid naming conflicts with skolem symbols, anonymous updates, etc.