Minor clean-up and smaller fixes.
  The NumericQuantifierTerm implementation binds now the variables as thought by the TermFactory and
the NumericQuantifer arity matches the expected arity of the TermFactory. But I am not sure if the expected behaviour
of the Termfactory is correct as I do not know synatx and semantics of \sum and \product. Maybe someone can look over it
and change it if necessary. Nevertheless the implementation is now (hopefully) consistent.