(Termporary) Reintegrated old Simplify translation
Due to the fact, that the old translation for Simplify seems to be more
reliable in concerns of the Test Generation feature of KeY (try the
binary tree example) we decided to adapt and include it to the current code.
Hopefully this is only a temporary solution as Simon is already working
on the refinement of the new translation.