Refactoring and new features in TestGeneration-Code
Refactoring:
fileWriter-flag in PrettyPrinter was removed. Appropriate Methods in
CompilableJavaPP respectivly CompilableJavaCardPP are used instead.

Modularisation of TestGeneration-Code was improved to ease the
implementation of differen behaviours.

New Feature:
While generation a TestSuite in KeY one is able to choose between 'Java'
or 'JavaCard'.

'JavaCard' is the old way where the CUT is replaced by a CUT that
contains everything from the original CUT plus getter and setter methods
for private fields.

'Java' is the newly implemented way where a helper-class is generated
that uses reflection to access the private fields.