Initialize MainWindow title from within MainWindow only
First of all MainWindow::initialize should do all the initializations
needed (including the Frame title) and in the second place a common
title string resource should still be maintained for "other projects"
(see Javadoc of removed getMainWindowTitle). And in the last place it
will be possible to both create and retrieve the MainWindow singleton by
calling only MainWindow::getInstance, when no argument is needed anymore
to instantiate the singleton.

- remove title argument from both MainWindow::createInstance and
  MainWindow::initialize
- add getUserInterfaceTitle method to KeYResourceManager so that every
  UserInterface can use a common title when it requires one
- remove getMainWindowTitle method from Main and set Frame title in
  MainWindow::initialize to what
  KeYResourceManager::getUserInterfaceTitle returns