Fix #1183: MainWindow does not implement Singleton correctly
Implement getUserinterface as Javadoc states and implicitly create the
singleton instance if it does not exist yet. Therefore createInstance
and hasInstance (I believe this existed to not risk throwing an
exception when calling getInstance without having called createInstance
before) are not needed anymore. Also allow access to the singleton
instance from within MainWindow solely to guarantee a single instance of
MainWindow only and declare the field private.

- remove both createInstance and hasInstance from MainWindow
- declare field MainWindow::instance private
- remove getUserInterface(boolean) from MainWindow as it does exactly
  the same as getUserInterface does currently and for the single call
  from Main a convenience implementation is not necessary I think