Add fieldReference factory to KeYJavaASTFactory
fieldReference takes Services and ExecutionContext to retrieve an
Expression's KeYJavaType. Afterwards the factory uses the Services to
select the field that is named as indicated by the caller from the just
determined class type. Then fieldReference is able to create the
FieldReference after it has turned the Expression into an
ParenthesizedExpression. Here the Expression under consideration is the
object on which the field is to be accessed.