[Taclets and Strategy] Support for trigger definition within taclets added
First support for the definition of triggers inside taclets.

A trigger is specified as follows

  \trigger{<triggerVar>} triggerExp (\avoid <cond1>, ..., <condN>);

The <triggerVar> is the schemavariable to be instantiated by the trigger.
The <condI> are atoms with comparator top level operator (<=, ==, >=). If one of the condI can be proven, the trigger will not be fired.

Limitations:
  * No trigger sets and set of triggers are yet supported
  * Taclet must contain only one uninstantiated variable
  * Need better condition prover