Fix taclet2Builder is null in taclet parsers
When the filename argument was removed from the `KeYParser`
constructors, the two constructors

    KeYParser(ParserMode,         TokenStream, Services, NamespaceSet)
    KeYParser(ParserMode, String, TokenStream, Services, NamespaceSet)

were left with the same signature. However, these two constructors had
different semantics with regard to the instanced `JavaReader`. Since
those semantics cannot be reflected in the constructor signature, use
the implementation from the version with the filename argument and
create the other `JavaReader` instance explicitly where needed. This
affects only the test suite `TestClashFreeSubst`.