[Fix] Hashcode computation of labeled term
The hash code computation of term was sensitive for the order of the labels while the equals method is not.
One problem remains, but that is most likely more a documentation problem, namely that neither of the two methods would work if we allow equal labels to occur several times.